requirements.txt
pandas 
yfinance
psycopg2-binary
python-dotenv